requirements.txt
asgiref
Django
djangorestframework
psycopg2-binary
sqlparse
drf-yasg
djangorestframework-simplejwt[crypto]
